Dev-Tools: initial push of some useful tools
- instructions.xml: is an xml file containing the description of
all the instructions decoded (or that should be) by the vARCH, with
code executed by those instructions.

- instr_parser.php: an initial generator script, which by now creates
a std_instructions.h header file. In future will produce an analogue file
for the asm-parser, too. This will help in mantaining correlation between
enanchements in the vARCH and in code geerated by the ASM-Parser.

- asm_parser.php: an initial generator script, that generates a lex file
containing rules for lexical parsing (and decoding) of asm sources.
(to be used with flex/lex)