vAsm-2.0: Initial commit of vAsm-2.0 which adapts to the changes in vArch-2.0
NOTE: It compiles fine, parses the code, and generates the AST tree, but fails to generate ASM.
The only usable parser is the Low parser (Mid has been disabled for now, but some features have been ported to Low).

The prominent features are exactly the ones described for vArch-2.0: it just adapts to architectural changes.

Signed-off-by: Nicola <ben.vighy@gmail.com>